kill git grep process when alarm triggered
Case: GH #50

There was already an alarm set aroun the "git run" command
but nothing was killing the "git" command running behind.

When the alarm is triggered check if we have a running git
command, if so, try to kill it.

Also guarantee that the alarm timer is always set (0 is invalid)