Fix output parser to show more results
Fix #63

The git output parser was pretty naive and buggy when multiple
chunks were found from the same file.

This was resulting in skipping results we should have displayed to begin with.
That changeset is fixing that issue and will now show more results.

The summary was accurate but the output truncated due to poor man parsing reason.